requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-environ>=0.11,<1.0
drf-spectacular>=0.27,<1.0
